Lower wdata_ready signal for one cycle after wlast goes high in simulation.
REASON: When consecutive multi-burst write commands are issued, wdata
is seen by simulator without a cycle delay, but the command requires a
cycle for ready to go high. While this is not a problem on boards, this
violates the simulator's view of which command is seen first. This
change forces the simulator to deassert wdata_ready after wlast, thereby
enforcing a cycle cycle separation between wdatas of two consecutive
bursts.